ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
scikit-learn==1.4.0
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.5
